+++ packages/lib/src/types.ts
@@ -2,7 +2,11 @@ import { InputParametersObject, OutputParametersObject } from '@connery-io/sdk';
 
 export type ActionOutput = {
   output: OutputParametersObject;
-  usedConnectorKey: string;
-  usedActionKey: string;
-  usedInputParameters: InputParametersObject;
+  used: UsedToRunAction;
+};
+
+export type UsedToRunAction = {
+  pluginKey: string;
+  actionKey: string;
+  inputParameters: InputParametersObject;
 };